do_vfs_ioctl(): use saner types
casting to pointer to int, only to pass that to function that
takes pointer to void and uses it as pointer to structure is
really asking for trouble.

"Some pointer, I'm not sure what to" is spelled "void *",
not "int *"; use that.

And declare the functions we are passing that pointer to
as taking the pointer to what they really want to access.

Signed-off-by: Al Viro <viro@zeniv.linux.org.uk>
Signed-off-by: Arnd Bergmann <arnd@arndb.de>
Signed-off-by: Divyanshu-Modi <divyan.m05@gmail.com>
Change-Id: If05ffe80d5ee0f5847dc62dc71e1a1c134064e35
Signed-off-by: Divyanshu-Modi <divyan.m05@gmail.com>