msm: kgsl: Add a setproperty option for target specific code
Pass setproperty through to target specific code to allow it to handle
KGSL_PROP_PWRCTRL on a per-target basis.

It is a bit more overkill to do this in a generic fashion instead of
defining a ->set_pwrctrl function but this gives more flexbility to
handle other target-specific properties in the future.

Change-Id: Ic0dedbade161c842262d36eb0cfc1410fa1fa774
Signed-off-by: Jordan Crouse <jcrouse@codeaurora.org>
Signed-off-by: Divyanshu-Modi <divyan.m05@gmail.com>