arm64: vendor/dts: Yupik: Remove display ramdump memory region
This region is the same as the early splash framebuffer initialized by
XBL. msm_drm uses it for ramdump when available, but that's a debugging
feature we don't need. It also wastes memory after early boot because
msm_drm will keep it reserved rather than freeing the framebuffer memory
once it has taken control of the display hardware.

Remove the display ramdump region to save 8 MiB of memory

Signed-off-by: Danny Lin <danny@kdrag0n.dev>
Signed-off-by: Kazuki Hashimoto <kazukih@tuta.io>
Signed-off-by: Divyanshu-Modi <divyan.m05@gmail.com>
Change-Id: I7409281b4b96caa278a59f729ff47be4dede103d
Signed-off-by: Divyanshu-Modi <divyan.m05@gmail.com>