interconnect: qcom: Set floor of 1 for BCM unit/width
Some BCMs that aren't directly associated with NOC ports (i.e. ACV)
publish their width as 0 in cmd_db. The client votes are multiplied by
this width when adjusting for node/BCM width differences. This means a
zero width also zeroes the client votes, making it impossible to vote on
these zero-width BCMs. Set a width floor of 1 to allow aggregation on
these nodes.

Also set a floor on the BCM units. They should never by 0 in cmd_db, but
if they are it would cause a div-by-zero.

Change-Id: I14ba89411ec02077db6d2c27b6de932c5b7dfe1e
Signed-off-by: Mike Tipton <mdtipton@codeaurora.org>
Signed-off-by: Divyanshu-Modi <divyan.m05@gmail.com>