Arm64: dts: msm: Use interconnect DT property to get icc handle
Make use of interconnects property along with of_icc_get() API.
This is mainly to prevent multiple deferred probe when geni tries
to probe and ICC is not probed.

This way the path from QUP SE to DDR will be from aggre1_noc ->
system_noc -> gem_noc -> mc_virt. Hence we need to define overall
3 nodes for interconnect path to DDR.

This change replaces existing ICC path vectors with the interconnect
path property. Respective driver change is taken cared and it depends
on this DTSI change.

Change-Id: I1944b285e49aa24d032b1d7481e0f8d2fe2e9913
Signed-off-by: Divyanshu-Modi <divyan.m05@gmail.com>