msm: kgsl: Use non-atomic snapshot memory if it is unused
Over time, memory gets fragmented, and eventually, a system that is
up for a long time may have only a few physically contiguous pages.
Hence, there is a chance of atomic snapshot memory allocation failure.
Use non-atomic snapshot memory if it is unused as it is allocated
during probe time.

Change-Id: I725adeb80614a775f92213ebe9768f4f2f1c70a2
Signed-off-by: Kamal Agrawal <kamaagra@codeaurora.org>
Signed-off-by: Divyanshu-Modi <divyan.m05@gmail.com>