power: supply: qti_battery_charger: add support for emergency shutdown
Currently, when SOC reaches 0 and device is not charging, we
depend on the userspace (e.g. Android battery manager service)
to initiate a shutdown. Sometimes, this takes more time than
expected (> 20 seconds). When the device is discharging with a
high system load, this may cause device to brown out as the
battery voltage can go below the safe operating level. Add
support to initiate an emergency shutdown when the following
conditions are met.

- Battery SOC is 0
- Battery is not charging
- Battery voltage is lower than a specified level

Add "qcom,shutdown-voltage" to support this.

Change-Id: I542ae08037c4e9db23d338b8de2eeb32c809a0a6
Signed-off-by: Subbaraman Narayanamurthy <subbaram@codeaurora.org>