ARM: dts: qcom: Add of version register to figure out ntn3 version
This change is to add version register to get the ntn3 switch version.
Added two separate reg update sequence ,
one is reg update this is needed to keep NTN3v1 in gen3,
this sequence is not needed in NTN3v2 as it is already taken care in v2
and another one is switch reg update this will update the
gen3 preset value to both NTN3v1 and NTN3v2.

Also added a optional force i2c settings flag to update the
demphasis setting prior pcie linkup in dtsi binding irrespective
of ntn3 chip version.

Change-Id: Ibb249af5455c76e789931962f2cb9449d8814a42