arm64: Inline the spinlock function family regardless of preemption
Combined with LTO, this yields a consistent 5% boost to procfs I/O
performance right off the bat (as measured with callbench). The spin
lock functions constitute some of the hottest code paths in the kernel

inlining them to improve performance makes sense.

back in 2005 due to possible instability and debugging requirements [1]
however this no longer applies. Inling spinlocks provides a performance
improvement without reducing the kernel's stability, even when preemption
is enabled.

Futhermore, while it's true that preemption increases the size of the
spinlock code (which in turn causes it to be less likely to hit the
cache) inlining the spinlock function family still improves performance

[1] : https://lore.kernel.org/lkml/20200530050943.142128-1-sultan@kerneltoast.com

Co-authored-by: Diab Neiroukh <lazerl0rd@thezest.dev>
Co-authored-by: Sultan Alsawaf <sultan@kerneltoast.com>
Change-Id: I70fb6f4696af8e772b8b29ba2d8dde1413a0b664
Signed-off-by: Divyanshu-Modi <divyan.m05@gmail.com>