power: supply: qti_battery_charger: Register wls_psy based on DT property
Currently, wireless power supply (wls_psy) is registered
by default to support wireless charging. However,
on some boards, wireless charging is not supported.
When the properties under wls_psy are read from the
userspace or a client, it times out and prints error
logs unnecessarily leading to a poor user experience.

Hence, add a DT property "qcom,wireless-not-supported"
which if specified wouldn't register the wls_psy.

Change-Id: If4ed36957fa8425018f80dfa43d82ee746a45566
Signed-off-by: Kamal Wadhwa <quic_kamalw@quicinc.com>
Signed-off-by: Divyanshu-Modi <divyan.m05@gmail.com>