slimbus: qcom-ngd-ctrl: Handle race condition between SSR and qmi request
A race issue occurs when the qmi_handle is released from the
ngd down sequence and  the same handle is utilised in the
qmi power on request, resulting in a crash.

Locking mechanism is added to make sure that only valid qmi
handles is being accessed during qmi power on request.

Change-Id: Ic519a927ec0fbc5d07c9b1aad433973478fe3d6d
Signed-off-by: Mehul Raninga <quic_mraninga@quicinc.com>
Signed-off-by: Divyanshu-Modi <divyan.m05@gmail.com>