qti_battery_charger: Add support for SCOPE prop in batttery psy
SCOPE is quered when power_supply_is_system_supplied is
called and apparently our touch panels call this func
for setting charging mode for the panels.

But, it ends up querying all the psy registered and spams
logcat when it cannot find the prop.

return value is initialised by -ENODATA in get_prop for
battery psy just return when SCOPE prop is called this
silences log and doesn't affect behaviour.

Test: Device boots and touch ic set the charge prop when charger is connected.
Change-Id: Ic00be9d7a0c3b55b4211a686bb8d2e78bc8712f2
Signed-off-by: Divyanshu-Modi <divyan.m05@gmail.com>