techpack: touch: Implement goodix_get_charging_status() for goodix_brl
This removes the dependency on the power_supply_is_system_supplied
which passes the SCOPE prop to the charger driver causing a logspam.

One fix was to add prop handling the qti_battery_charger driver itself
but, fixing the root cause is much more concise as the charger mode
has to be set only when the charger is connected and eventually the
above mentioned function does the same but, spams the SCOPE prop
and is prone to change due to upstrean changes down the line.

Maintain the functionality and fix the problem by querying USB_ONLINE
prop.

Test: Boots and charger mode prop is passed on usb connect

Change-Id: I21a056546ed74a85f3ddbc7bcf59bb7b35a97ed7
Signed-off-by: Divyanshu-Modi <divyan.m05@gmail.com>